# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: DealFlow360 API
  description: Comprehensive quotation-to-cash platform API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
paths:
...
  /inventory/warehouses:
    get:
      summary: List warehouses
      tags:
        - Inventory
      responses:
        '200':
          description: Warehouse list
    post:
      summary: Create warehouse
      tags:
        - Inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
                leadTimeDays:
                  type: number
                shippingCostWeight:
                  type: number
      responses:
        '201':
          description: Warehouse created
  /inventory/warehouses/{warehouseId}:
    patch:
      summary: Update warehouse
      tags:
        - Inventory
      parameters:
        - name: warehouseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
                active:
                  type: boolean
      responses:
        '200':
          description: Warehouse updated
  /inventory/warehouses/{warehouseId}/balances:
    get:
      summary: List inventory balances
      tags:
        - Inventory
      parameters:
        - name: warehouseId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Inventory balances
  /inventory/warehouses/{warehouseId}/adjust:
    post:
      summary: Adjust inventory
      tags:
        - Inventory
      parameters:
        - name: warehouseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                variantId:
                  type: string
                quantity:
                  type: string
                reason:
                  type: string
      responses:
        '200':
          description: Inventory adjusted
...
  /inventory/stock-movements/receipt:
    post:
      summary: Create stock receipt
      tags:
        - Inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                warehouseId:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                      variantId:
                        type: string
                      quantity:
                        type: string
                reference:
                  type: string
      responses:
        '201':
          description: Receipt created
  /fulfillment/orders/{orderId}/fulfillment/preview:
    get:
      summary: Preview fulfillment allocation
      tags:
        - Fulfillment
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fulfillment preview
  /fulfillment/orders/{orderId}/fulfillment/reserve:
    post:
      summary: Reserve stock for fulfillment
      tags:
        - Fulfillment
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                planId:
                  type: string
      responses:
        '200':
          description: Stock reserved
  /fulfillment/orders/{orderId}/fulfillment/override:
    post:
      summary: Override fulfillment allocation
      tags:
        - Fulfillment
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                allocations:
                  type: array
                  items:
                    type: object
                    properties:
                      warehouseId:
                        type: string
                      quoteLineId:
                        type: string
                      quantity:
                        type: string
                reason:
                  type: string
      responses:
        '200':
          description: Allocation overridden
  /fulfillment/orders/{orderId}/shipments:
    get:
      summary: List shipments for order
      tags:
        - Fulfillment
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Shipments list
  /fulfillment/shipments/{shipmentId}/ship:
    post:
      summary: Ship shipment
      tags:
        - Fulfillment
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                trackingNumber:
                  type: string
      responses:
        '200':
          description: Shipment shipped
...
  /fulfillment/backorders/{backorderId}/consolidate:
    post:
      summary: Consolidate backorder
      tags:
        - Fulfillment
      parameters:
        - name: backorderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Backorder consolidated
...
  /billing/invoices/{invoiceId}:
    get:
      summary: Get invoice
      tags:
        - Billing
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice details
  /billing/invoices/{invoiceId}/issue:
    post:
      summary: Issue invoice
      tags:
        - Billing
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice issued
  /billing/invoices/{invoiceId}/payments:
    get:
      summary: List invoice payments
      tags:
        - Billing
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payments list
    post:
      summary: Record payment
      tags:
        - Billing
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: string
                method:
                  type: string
                  enum: [BANK_TRANSFER, CREDIT_CARD, CHECK, OTHER]
                reference:
                  type: string
                paymentDate:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Payment recorded
...
  /billing/credit-notes/{creditNoteId}/apply:
    post:
      summary: Apply credit note to invoice
      tags:
        - Billing
      parameters:
        - name: creditNoteId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                invoiceId:
                  type: string
      responses:
        '200':
          description: Credit note applied
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session